Data.Serde.Internal.ISyn

Description

intermediate syntax for the parser

see Data.Serde.Syn for the grammar

Documentation

parsetop :: Parser (Derive, [ISyn])

data ISyn

intermediate representation of type definitions parsed from the input syntax. can be a data type, newtype, or type alias.

Constructors

ISynData

Fields

isnam :: String

isflds :: [ISynFld]

ISynNewtype

Fields

isnam :: String

isfld1 :: INewtypePayload

ISynAlias

Fields

isnam :: String

isdest :: String

data ISynFld

field definition for data types and newtypes, containing a field name and its type information

Constructors

ISynFld

Fields

isfldnam :: String

isfldtyp :: ViaInfo

data INewtypePayload

payload for newtype definitions, which can either be a field with a name and type, or just a type

Constructors

IField ISynFld

IType ViaInfo

data ViaInfo

type information that may include a via clause

Constructors

Plain String

WithVia String String

newtype Derive

collection of type classes to derive for all defined types

Constructors

Derive

Fields

getderive :: [String]

type Parser = Parsec Void String

parser type for the intermediate syntax

debug

_testbody1 :: String